example/test/pyproject.toml
[project]
name = "my_mcp_project"
version = "0.1.0"
description = "Auto-generated MCP project"
authors = [
    {name="YourName", email="you@example.com"}
]
dependencies = [
    "fastmcp>=2.3.3",
    "starlette>=0.25.0",
    "uvicorn>=0.22.0",
    "viyv_mcp>=0.1.2",
    "openai-agents>=0.0.13",
    "pydantic>=2",
    "slack-bolt>=1.23.0",
    "aiohttp>=3.11.18",
    "pillow>=11.2.1",
    "claude-code-sdk>=0.0.11",
    "httpx>=0.28.1",
]
